feat: add battle state debug observer
